Adds support for line continuation via indentation
Allows long blocks of text or arrays to be cleaning added to an ini file.
If a line begins with empty space (tab, space) it will automatically
be appended to the previous line.

! This assumes that section heads and keys are NOT indented. !

Example ini:

[SectionHead]
key1=value of key 1
my_array=val1,
         val2,
	 val3
long_text=This is some long text
          that I need to include
	  it wraps multiple lines.

The resulting object:

{SectionHead:
 {
  key1:'value of key 1',
  my_array:'val1, val2, val3',
  long_text:'This is some long text that I need to include it wraps multiple lines.'
 }
}